Move GatewayConfig validation configs to PaymentBundle
