+++ src/Sylius/Bundle/PaymentBundle/Resources/config/app/config.yml
@@ -7,13 +7,3 @@ jms_serializer:
             sylius-payment:
                 namespace_prefix: "Sylius\\Component\\Payment"
                 path: "@SyliusPaymentBundle/Resources/config/serializer"
-
-sylius_payment:
-    gateway_config:
-        validation_groups:
-            paypal_express_checkout:
-                - 'sylius'
-                - 'sylius_paypal_express_checkout'
-            stripe_checkout:
-                - 'sylius'
-                - 'sylius_stripe_checkout'
