+++ src/Sylius/Bundle/PaymentBundle/Resources/config/services.xml
@@ -13,7 +13,7 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <imports>
-        <import resource="services/form.xml" />
+        <import resource="services/**/*.xml" />
     </imports>
 
     <parameters>
