+++ src/Sylius/Bundle/PaymentBundle/Resources/config/validation/GatewayConfig.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping"
+                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                    xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd"
+>
+    <class name="Sylius\Component\Payment\Model\GatewayConfig">
+        <property name="gatewayName">
+            <constraint name="NotBlank">
+                <option name="message">sylius.gateway_config.gateway_name.not_blank</option>
+                <option name="groups">sylius</option>
+            </constraint>
+        </property>
+
+        <property name="factoryName">
+            <constraint name="NotBlank">
+                <option name="message">sylius.gateway_config.factory_name.not_blank</option>
+                <option name="groups">sylius</option>
+            </constraint>
+
+            <constraint name="Sylius\Bundle\PaymentBundle\Validator\Constraints\GatewayFactoryExists">
+                <option name="groups">
+                    <value>sylius</value>
+                </option>
+            </constraint>
+        </property>
+    </class>
+</constraint-mapping>
