+++ src/Sylius/Bundle/PaymentBundle/Validator/GatewayFactoryExistsValidator.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\PayumBundle\Validator;
+namespace Sylius\Bundle\PaymentBundle\Validator;
 
-use Sylius\Bundle\PayumBundle\Validator\Constraints\GatewayFactoryExists;
+use Sylius\Bundle\PaymentBundle\Validator\Constraints\GatewayFactoryExists;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
