+++ src/Sylius/Bundle/PayumBundle/Resources/translations/validators.en.yml
@@ -1,10 +1,5 @@
 sylius:
     gateway_config:
-        invalid_gateway_factory: Invalid gateway factory. Available factories are {{ available_factories }}.
-        gateway_name:
-            not_blank: Please enter gateway name.
-        factory_name:
-            not_blank: Please enter gateway factory name.
         paypal:
             password:
                 not_blank: Please enter paypal password.
